+++ app/src/main/java/edu/ucne/registrotecnicos/MainActivity.kt
@@ -244,26 +244,5 @@ class MainActivity : ComponentActivity() {
         val sueldo: Double = 0.0,
     )
 
-    @Dao
-    interface TecnicosDao {
-        @Upsert()
-        suspend fun save(tecnico: TecnicosEntity)
-
-        @Query(
-            """
-            Select *
-            from Tecnicos
-            where tecnicosId = :id
-            limit 1
-            """
-        )
-        suspend fun find(id: Int): TecnicosEntity?
-
-        @Delete()
-        suspend fun delete(tecnicosEntity: TecnicosEntity)
-
-        @Query("Select * from Tecnicos")
-        fun getAll(): Flow<List<TecnicosEntity>>
-    }
 
 }
